+++ src/docs/SCR_NotFound.svelte
@@ -1,0 +1,57 @@
+<script>
+  import routerStore from "../js/store/router.js";
+  import { SCR_ROUTER_LINK } from "../index.js";
+
+  let backTo = "/rootRoute";
+  $: if ($routerStore.currentRoute) {
+    backTo = $routerStore.currentRoute.name;
+  }
+</script>
+
+<center>
+  <p class="scr-p">Not Found</p>
+  <p class="scr-p-small">{$routerStore.currentLocation || "='("}</p>
+  <br />
+  <SCR_ROUTER_LINK to={{ name: backTo }}>
+    <div class="scr-menu-item">Back</div>
+  </SCR_ROUTER_LINK>
+</center>
+
+<style>
+  .scr-p {
+    color: #ff3e00;
+    text-transform: uppercase;
+    font-size: 4rem;
+    font-weight: 100;
+  }
+
+  .scr-p-small {
+    color: #ff3e00;
+    text-transform: uppercase;
+    font-size: 2rem;
+    font-weight: 100;
+  }
+
+  .scr-menu-item {
+    width: 50% !important;
+    display: grid;
+    grid-template-columns: 1fr;
+    justify-content: left;
+    align-content: center;
+    width: 100%;
+    padding: 0.5rem;
+    background-color: #ff440009;
+    border-radius: 0.5rem;
+    transition: all linear 0.2s;
+    text-align: center;
+    color: #a82d00;
+    margin-bottom: 0.5rem;
+  }
+
+  .scr-menu-item:hover {
+    background-color: #ff440018;
+    cursor: pointer;
+    font-weight: bolder;
+    color: #d43900;
+  }
+</style>